+++ src/resources/workflow/workflow.service.js
@@ -1,0 +1,378 @@
+import teamController from '../team/team.controller';
+import constants from '../utilities/constants.util';
+import emailGenerator from '../utilities/emailGenerator.util';
+import notificationBuilder from '../utilities/notificationBuilder';
+
+import moment from 'moment';
+import _ from 'lodash';
+
+export default class WorkflowService {
+	constructor(workflowRepository) {
+		this.workflowRepository = workflowRepository;
+	}
+
+	async createNotifications(context, type = '') {
+		if (!_.isEmpty(type)) {
+			// local variables set here
+			let custodianManagers = [],
+				managerUserIds = [],
+				options = {},
+				html = '';
+
+			// deconstruct context
+			let { publisherObj, workflow = {}, actioner = '' } = context;
+
+			// switch over types
+			switch (type) {
+				case constants.notificationTypes.WORKFLOWCREATED:
+					// 1. Get managers for publisher
+					custodianManagers = teamController.getTeamMembersByRole(publisherObj, constants.roleTypes.MANAGER);
+					// 2. Get managerIds for notifications
+					managerUserIds = custodianManagers.map(user => user.id);
+					// 3. deconstruct workflow
+					let { workflowName = 'Workflow Title', _id, steps, createdAt } = workflow;
+					// 4. setup options
+					options = {
+						actioner,
+						workflowName,
+						_id,
+						steps,
+						createdAt,
+					};
+					// 4. Create notifications for the managers only
+					await notificationBuilder.triggerNotificationMessage(
+						managerUserIds,
+						`A new workflow of ${workflowName} has been created`,
+						'workflow',
+						_id
+					);
+					// 5. Generate the email
+					html = await emailGenerator.generateWorkflowCreated(options);
+					// 6. Send email to custodian managers only within the team
+					await emailGenerator.sendEmail(custodianManagers, constants.hdrukEmail, `A Workflow has been created`, html, false);
+					break;
+			}
+		}
+	}
+
+	calculateStepDeadlineReminderDate(step) {
+		// Extract deadline and reminder offset in days from step definition
+		let { deadline, reminderOffset } = step;
+		// Subtract SLA reminder offset
+		let reminderPeriod = +deadline - +reminderOffset;
+		return `P${reminderPeriod}D`;
+	}
+
+	buildNextStep(userId, application, activeStepIndex, override) {
+		// Check the current position of the application within its assigned workflow
+		const finalStep = activeStepIndex === application.workflow.steps.length - 1;
+		const requiredReviews = application.workflow.steps[activeStepIndex].reviewers.length;
+		const completedReviews = application.workflow.steps[activeStepIndex].recommendations.length;
+		const stepComplete = completedReviews === requiredReviews;
+		// Establish base payload for Camunda
+		// (1) phaseApproved is passed as true when the manager is overriding the current step/phase
+		//		this short circuits the review process in the workflow and closes any remaining user tasks
+		//		i.e. reviewers within the active step OR when the last reviewer in the step submits a vote
+		// (2) managerApproved is passed as true when the manager is approving the entire application
+		//		from any point within the review process
+		// (3) finalPhaseApproved is passed as true when the final step is completed naturally through all
+		//		reviewers casting their votes
+		let bpmContext = {
+			businessKey: application._id,
+			dataRequestUserId: userId.toString(),
+			managerApproved: override,
+			phaseApproved: (override && !finalStep) || stepComplete,
+			finalPhaseApproved: finalStep,
+			stepComplete,
+		};
+		if (!finalStep) {
+			// Extract the information for the next step defintion
+			let { name: dataRequestPublisher } = application.publisherObj;
+			let nextStep = application.workflow.steps[activeStepIndex + 1];
+			let reviewerList = nextStep.reviewers.map(reviewer => reviewer._id.toString());
+			let { stepName: dataRequestStepName } = nextStep;
+			// Update Camunda payload with the next step information
+			bpmContext = {
+				...bpmContext,
+				dataRequestPublisher,
+				dataRequestStepName,
+				notifyReviewerSLA: this.calculateStepDeadlineReminderDate(nextStep),
+				reviewerList,
+			};
+		}
+		return bpmContext;
+	}
+
+	getWorkflowCompleted(workflow = {}) {
+		let workflowCompleted = false;
+		if (!_.isEmpty(workflow)) {
+			let { steps } = workflow;
+			workflowCompleted = steps.every(step => step.completed);
+		}
+		return workflowCompleted;
+	}
+
+	getActiveWorkflowStep(workflow = {}) {
+		let activeStep = {};
+		if (!_.isEmpty(workflow)) {
+			let { steps } = workflow;
+			activeStep = steps.find(step => {
+				return step.active;
+			});
+		}
+		return activeStep;
+	}
+
+	getStepReviewers(step = {}) {
+		let stepReviewers = [];
+		// Attempt to get step reviewers if workflow passed
+		if (!_.isEmpty(step)) {
+			// Get active reviewers
+			if (step) {
+				({ reviewers: stepReviewers } = step);
+			}
+		}
+		return stepReviewers;
+	}
+
+	getRemainingReviewers(Step = {}, users) {
+		let { reviewers = [], recommendations = [] } = Step;
+		let remainingActioners = reviewers.filter(
+			reviewer => !recommendations.some(rec => rec.reviewer.toString() === reviewer._id.toString())
+		);
+		remainingActioners = [...users].filter(user => remainingActioners.some(actioner => actioner._id.toString() === user._id.toString()));
+
+		return remainingActioners;
+	}
+
+	getActiveStepStatus(activeStep, users = [], userId = '') {
+		let reviewStatus = '',
+			deadlinePassed = false,
+			remainingActioners = [],
+			decisionMade = false,
+			decisionComments = '',
+			decisionApproved = false,
+			decisionDate = '',
+			decisionStatus = '';
+		let { stepName, deadline, startDateTime, reviewers = [], recommendations = [], sections = [] } = activeStep;
+		let deadlineDate = moment(startDateTime).add(deadline, 'days');
+		let diff = parseInt(deadlineDate.diff(new Date(), 'days'));
+		if (diff > 0) {
+			reviewStatus = `Deadline in ${diff} days`;
+		} else if (diff < 0) {
+			reviewStatus = `Deadline was ${Math.abs(diff)} days ago`;
+			deadlinePassed = true;
+		} else {
+			reviewStatus = `Deadline is today`;
+		}
+		remainingActioners = reviewers.filter(reviewer => !recommendations.some(rec => rec.reviewer.toString() === reviewer._id.toString()));
+		remainingActioners = users
+			.filter(user => remainingActioners.some(actioner => actioner._id.toString() === user._id.toString()))
+			.map(user => {
+				let isCurrentUser = user._id.toString() === userId.toString();
+				return `${user.firstname} ${user.lastname}${isCurrentUser ? ` (you)` : ``}`;
+			});
+
+		let isReviewer = reviewers.some(reviewer => reviewer._id.toString() === userId.toString());
+		let hasRecommended = recommendations.some(rec => rec.reviewer.toString() === userId.toString());
+
+		decisionMade = isReviewer && hasRecommended;
+
+		if (decisionMade) {
+			decisionStatus = 'Decision made for this phase';
+		} else if (isReviewer) {
+			decisionStatus = 'Decision required';
+		} else {
+			decisionStatus = '';
+		}
+
+		if (hasRecommended) {
+			let recommendation = recommendations.find(rec => rec.reviewer.toString() === userId.toString());
+			({ comments: decisionComments, approved: decisionApproved, createdDate: decisionDate } = recommendation);
+		}
+
+		let reviewPanels = sections.map(section => constants.darPanelMapper[section]).join(', ');
+
+		return {
+			stepName,
+			remainingActioners: remainingActioners.join(', '),
+			deadlinePassed,
+			isReviewer,
+			reviewStatus,
+			decisionMade,
+			decisionApproved,
+			decisionDate,
+			decisionStatus,
+			decisionComments,
+			reviewPanels,
+		};
+	}
+
+	getWorkflowStatus(application) {
+		let workflowStatus = {};
+		let { workflow = {} } = application;
+		if (!_.isEmpty(workflow)) {
+			let { workflowName, steps } = workflow;
+			// Find the active step in steps
+			let activeStep = this.getActiveWorkflowStep(workflow);
+			let activeStepIndex = steps.findIndex(step => {
+				return step.active === true;
+			});
+			if (activeStep) {
+				let { reviewStatus, deadlinePassed } = this.getActiveStepStatus(activeStep);
+				//Update active step with review status
+				steps[activeStepIndex] = {
+					...steps[activeStepIndex],
+					reviewStatus,
+					deadlinePassed,
+				};
+			}
+			//Update steps with user friendly review sections
+			let formattedSteps = [...steps].reduce((arr, item) => {
+				let step = {
+					...item,
+					sections: [...item.sections].map(section => constants.darPanelMapper[section]),
+				};
+				arr.push(step);
+				return arr;
+			}, []);
+
+			workflowStatus = {
+				workflowName,
+				steps: formattedSteps,
+				isCompleted: this.getWorkflowCompleted(workflow),
+			};
+		}
+		return workflowStatus;
+	}
+
+	getReviewStatus(application, userId) {
+		let inReviewMode = false,
+			reviewSections = [],
+			isActiveStepReviewer = false,
+			hasRecommended = false;
+		// Get current application status
+		let { applicationStatus } = application;
+		// Check if the current user is a reviewer on the current step of an attached workflow
+		let { workflow = {} } = application;
+		if (!_.isEmpty(workflow)) {
+			let { steps } = workflow;
+			let activeStep = steps.find(step => {
+				return step.active === true;
+			});
+			if (activeStep) {
+				isActiveStepReviewer = activeStep.reviewers.some(reviewer => reviewer._id.toString() === userId.toString());
+				reviewSections = [...activeStep.sections];
+
+				let { recommendations = [] } = activeStep;
+				if (!_.isEmpty(recommendations)) {
+					hasRecommended = recommendations.some(rec => rec.reviewer.toString() === userId.toString());
+				}
+			}
+		}
+		// Return active review mode if conditions apply
+		if (applicationStatus === 'inReview' && isActiveStepReviewer) {
+			inReviewMode = true;
+		}
+
+		return { inReviewMode, reviewSections, hasRecommended };
+	}
+
+	getWorkflowEmailContext(accessRecord, workflow, relatedStepIndex) {
+		// Extract workflow email variables
+		const { dateReviewStart = '' } = accessRecord;
+		const { workflowName, steps } = workflow;
+		const { stepName, startDateTime = '', endDateTime = '', completed = false, deadline: stepDeadline = 0, reminderOffset = 0 } = steps[
+			relatedStepIndex
+		];
+		const stepReviewers = this.getStepReviewers(steps[relatedStepIndex]);
+		const reviewerNames = [...stepReviewers].map(reviewer => `${reviewer.firstname} ${reviewer.lastname}`).join(', ');
+		const reviewSections = [...steps[relatedStepIndex].sections].map(section => constants.darPanelMapper[section]).join(', ');
+		const stepReviewerUserIds = [...stepReviewers].map(user => user.id);
+		const currentDeadline = stepDeadline === 0 ? 'No deadline specified' : moment().add(stepDeadline, 'days');
+		let nextStepName = '',
+			nextReviewerNames = '',
+			nextReviewSections = '',
+			duration = '',
+			totalDuration = '',
+			nextDeadline = '',
+			dateDeadline = '',
+			deadlineElapsed = false,
+			deadlineApproaching = false,
+			remainingReviewers = [],
+			remainingReviewerUserIds = [];
+
+		// Calculate duration for step if it is completed
+		if (completed) {
+			if (!_.isEmpty(startDateTime.toString()) && !_.isEmpty(endDateTime.toString())) {
+				duration = moment(endDateTime).diff(moment(startDateTime), 'days');
+				duration = duration === 0 ? `Same day` : duration === 1 ? `1 day` : `${duration} days`;
+			}
+		} else {
+			//If related step is not completed, check if deadline has elapsed or is approaching
+			if (!_.isEmpty(startDateTime.toString()) && stepDeadline != 0) {
+				dateDeadline = moment(startDateTime).add(stepDeadline, 'days');
+				deadlineElapsed = moment().isAfter(dateDeadline, 'second');
+
+				// If deadline is not elapsed, check if it is within SLA period
+				if (!deadlineElapsed && reminderOffset !== 0) {
+					let deadlineReminderDate = moment(dateDeadline).subtract(reminderOffset, 'days');
+					deadlineApproaching = moment().isAfter(deadlineReminderDate, 'second');
+				}
+			}
+			// Find reviewers of the current incomplete phase
+			let accessRecordObj = accessRecord.toObject();
+			if (_.has(accessRecordObj, 'publisherObj.team.users')) {
+				let {
+					publisherObj: {
+						team: { users = [] },
+					},
+				} = accessRecordObj;
+				remainingReviewers = getRemainingReviewers(steps[relatedStepIndex], users);
+				remainingReviewerUserIds = [...remainingReviewers].map(user => user.id);
+			}
+		}
+
+		// Check if there is another step after the current related step
+		if (relatedStepIndex + 1 === steps.length) {
+			// If workflow completed
+			nextStepName = 'No next step';
+			// Calculate total duration for workflow
+			if (steps[relatedStepIndex].completed && !_.isEmpty(dateReviewStart.toString())) {
+				totalDuration = moment().diff(moment(dateReviewStart), 'days');
+				totalDuration = totalDuration === 0 ? `Same day` : duration === 1 ? `1 day` : `${duration} days`;
+			}
+		} else {
+			// Get details of next step if this is not the final step
+			({ stepName: nextStepName } = steps[relatedStepIndex + 1]);
+			let nextStepReviewers = this.getStepReviewers(steps[relatedStepIndex + 1]);
+			nextReviewerNames = [...nextStepReviewers].map(reviewer => `${reviewer.firstname} ${reviewer.lastname}`).join(', ');
+			nextReviewSections = [...steps[relatedStepIndex + 1].sections].map(section => constants.darPanelMapper[section]).join(', ');
+			let { deadline = 0 } = steps[relatedStepIndex + 1];
+			nextDeadline = deadline === 0 ? 'No deadline specified' : moment().add(deadline, 'days');
+		}
+		return {
+			workflowName,
+			steps,
+			stepName,
+			startDateTime,
+			endDateTime,
+			stepReviewers,
+			duration,
+			totalDuration,
+			reviewerNames,
+			stepReviewerUserIds,
+			reviewSections,
+			currentDeadline,
+			nextStepName,
+			nextReviewerNames,
+			nextReviewSections,
+			nextDeadline,
+			dateDeadline,
+			deadlineElapsed,
+			deadlineApproaching,
+			remainingReviewers,
+			remainingReviewerUserIds,
+		};
+	}
+}
